pyproject.toml
[project]
name = "groq-arize"
version = "0.1.0"
description = "Groq + Arize Phoenix observability example"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arize-phoenix-otel>=0.12.1",
    "dotenv>=0.9.9",
    "groq>=0.28.0",
    "openinference-instrumentation-groq>=0.1.11",
    "python-dotenv>=1.0.0",
]
